cambio tipo long long
para poder manejar numeros mas grandes
implementaciones de hud en award y navegacion entre estados

+++ DeathTokenBase/DeathToken/Award.h
@@ -9,19 +9,19 @@ class Marbles;
 class Award : public GameState
 {
 private:
-	int betG;
-	int mWinG;
+	long long betG;
+	long long mWinG;
 	GameState* state;
 	Uint32 startTime;
 	Texture* background;
 	Text* text;
 	Text* winText;
-	int currentWin;
+	long long currentWin;
 	int wSize = 200;
 	int nSize = 300;
 	int cSize = 20;
 public:
-	Award(Game* game, GameState* lastState, int bet, int mWin);
+	Award(Game* game, GameState* lastState, long long bet, long long mWin);
 	void render()const  override;
 	void update() override;
 	std::string getWinMessage(int multiplier);
